requirements.txt
streamlit
faiss-cpu
pypdf
python-dotenv
tiktoken
openai>=1.40.0
